Moved session from core to the firenado root.
Refs: #109
